+++ bluebird/api/static/__init__.py
@@ -1,5 +1,0 @@
-"""
-Provides any static content for the application
-"""
-
-from .readme import readme