+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump required BlueSky version to `1.2.0`
 - When in `agent` mode, aircraft data is only logged after each `STEP` command
 
+### Removed
+
+- Removed static page which rendered and served the README markdown file
+- Removed the markdown pip dependency
+ 
+
 ## [1.2.1] - 2019-06-11
 
 ### Changed